+++ compiler/src/help.ml
@@ -9,6 +9,7 @@ let show_intrinsics asmOp fmt =
     | PrimX _ -> 2
     | PrimV _ -> 3
     | PrimVV _ -> 4
+    | PrimARM _ -> 5
   in
   let headers = [|
       "no size suffix";
@@ -16,8 +17,9 @@ let show_intrinsics asmOp fmt =
       "a zero/sign extend suffix, e.g., “_u32u16”";
       "one vector description suffix, e.g., “_4u64”";
       "two vector description suffixes, e.g., “_2u16_2u64”";
+      "a flag setting suffix (i.e. “S”) and a condition suffix (i.e. “cc”)"
     |] in
-  let intrinsics = Array.make 5 [] in
+  let intrinsics = Array.make 6 [] in
   List.iter (fun (n, i) ->
       let j = index i in
       intrinsics.(j) <- n :: intrinsics.(j))