+++ proofs/arch/arch_extra.v
@@ -3,7 +3,9 @@ From mathcomp Require Import all_ssreflect all_algebra.
 From mathcomp.word Require Import ssrZ.
 Require Import xseq strings utils var type values sopn expr arch_decl.
 Require Import compiler_util.
-Require Import sem_pexpr_params.
+Require Import
+  sem_pexpr_params
+  shift_kind.
 
 Set   Implicit Arguments.
 Unset Strict Implicit.
@@ -160,6 +162,7 @@ Definition sopn_prim_constructor (f:option wsize -> asm_op -> extended_op) (p : 
   | PrimSV x => sopn.PrimV (fun ws1 s v ws2 => f ws1 (x s v ws2))
   | PrimX x => sopn.PrimX (fun ws1 ws2 ws3 => f ws1 (x ws2 ws3))
   | PrimVV x => sopn.PrimVV (fun ws1 v1 ws2 v2 ws3 => f ws1 (x v1 ws2 v2 ws3))
+  | PrimARM x => sopn.PrimARM (fun sf ic hs => f None (x sf ic hs))
   end.
 (* duplication with lang/sopn.v -> maybe we can have one version to rule them all? *)
 Definition map_prim_constructor {A B} (f:A -> B) (p : sopn.prim_constructor A) :=
@@ -169,6 +172,7 @@ Definition map_prim_constructor {A B} (f:A -> B) (p : sopn.prim_constructor A) :
   | sopn.PrimV x => sopn.PrimV (fun ws1 s v ws2 => f (x ws1 s v ws2))
   | sopn.PrimX x => sopn.PrimX (fun ws1 ws2 ws3 => f (x ws1 ws2 ws3))
   | sopn.PrimVV x => sopn.PrimVV (fun ws1 v1 ws2 v2 ws3 => f (x ws1 v1 ws2 v2 ws3))
+  | sopn.PrimARM x => sopn.PrimARM (fun sf ic hs => f (x sf ic hs))
   end.
 Definition sopn_prim_string_base (o : seq (string * prim_constructor asm_op)) :=
   let to_ex ws o := BaseOp (ws, o) in