+++ proofs/lang/extraction.v
@@ -1,4 +1,5 @@
 Require Import var compiler.
+Require arm_params arm_sem.
 Require x86_params x86_sem.
 
 Require ExtrOcamlBasic.
@@ -31,6 +32,22 @@ Cd  "lang/ocaml".
 
 Extraction Blacklist String List Nat Utils Var Array.
 
-Separate Extraction utils sopn expr sem arch_decl x86_decl x86_sem.x86_prog x86_instr_decl x86_extra x86_params compiler.
+Separate Extraction
+  utils
+  sopn
+  expr
+  sem
+  arch_decl
+  x86_decl
+  x86_sem.x86_prog
+  x86_instr_decl
+  x86_extra
+  x86_params
+  arm_decl
+  arm_sem.arm_prog
+  arm_instr_decl
+  arm_extra
+  arm_params
+  compiler.
 
 Cd  "../..".